More fixes- 3.1 alpha tag and TicketsResource seperated split out.
