+++ appveyor.yml
@@ -1,5 +1,5 @@
 image: Visual Studio 2017
-version: 3.0.0-beta-{build}
+version: 3.1.0-alpha-{build}
 configuration: Release
 before_build:
   - dotnet restore
