+++ src/ZendeskApi.Client/Exceptions/ZendeskRequestExceptionBuilder.cs
@@ -3,6 +3,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
+using ZendeskApi.Client.Extensions;
 
 namespace ZendeskApi.Client.Exceptions
 {
