+++ src/ZendeskApi.Client/ZendeskClient.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using ZendeskApi.Client.Resources;
