+++ test/ZendeskApi.Client.Tests/Resources/SatisfactionRatingsTests.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -6,6 +6,7 @@ using Newtonsoft.Json;
 using Xunit;
 using ZendeskApi.Client.Models;
 using ZendeskApi.Client.Resources;
+using ZendeskApi.Client.Tests.ResourcesSampleSites;
 
 namespace ZendeskApi.Client.Tests.Resources
 {
