+++ test/ZendeskApi.Client.Tests/Resources/SearchResourceTests.cs
@@ -6,6 +6,7 @@ using Xunit;
 using ZendeskApi.Client.Models;
 using ZendeskApi.Client.Resources;
 using ZendeskApi.Client.Responses;
+using ZendeskApi.Client.Tests.ResourcesSampleSites;
 
 namespace ZendeskApi.Client.Tests.Resources
 {
