+++ test/ZendeskApi.Client.Tests/ResourcesSampleSites/SatisfactionRatingsResourceSampleSite.cs
@@ -9,12 +9,12 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
+using ZendeskApi.Client.Extensions;
 using ZendeskApi.Client.Models;
 using ZendeskApi.Client.Responses;
 using ZendeskApi.Client.Tests.Extensions;
-using ZendeskApi.Client.Tests.ResourcesSampleSites;
 
-namespace ZendeskApi.Client.Tests
+namespace ZendeskApi.Client.Tests.ResourcesSampleSites
 {
     public class SatisfactionRatingsResourceSampleSite : SampleSite
     {
