+++ test/ZendeskApi.Client.Tests/ResourcesSampleSites/SearchResourceSampleSite.cs
@@ -10,9 +10,8 @@ using Microsoft.Extensions.DependencyInjection;
 using ZendeskApi.Client.Models;
 using ZendeskApi.Client.Responses;
 using ZendeskApi.Client.Tests.Extensions;
-using ZendeskApi.Client.Tests.ResourcesSampleSites;
 
-namespace ZendeskApi.Client.Tests
+namespace ZendeskApi.Client.Tests.ResourcesSampleSites
 {
     public class SearchResourceSampleSite : SampleSite
     {
@@ -24,7 +23,7 @@ namespace ZendeskApi.Client.Tests
                     .MapGet("api/v2/search", (req, resp, routeData) =>
                     {
                         var obj = new ISearchResult[] {
-                            //new TicketResponse { Id = 1, Url = new Uri("https://company.zendesk.com/api/v2/tickets/1.json") },
+                            new TicketResponse { Id = 1, Url = new Uri("https://company.zendesk.com/api/v2/tickets/1.json") },
                             new Group { Id = 2, Url = new Uri("https://company.zendesk.com/api/v2/groups/2.json") },
                             new Organization { Id = 3, Url = new Uri("https://company.zendesk.com/api/v2/organizations/3.json") },
                             new User { Id = 4, Url = new Uri("https://company.zendesk.com/api/v2/users/4.json") }};
