+++ test/ZendeskApi.Client.Tests/ResourcesSampleSites/UserIdentitiesResourceSampleSite.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
+using ZendeskApi.Client.Extensions;
 using ZendeskApi.Client.Models;
 using ZendeskApi.Client.Responses;
 using ZendeskApi.Client.Tests.Extensions;
