+++ test/ZendeskApi.Client.Tests/ZendeskApiClientTests.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.Extensions.Options;
+using Microsoft.Extensions.Options;
 using Xunit;
 using ZendeskApi.Client.Options;
